started working on ResourceRepositoryin issue #32
